+++ src/com/inventory/db/repositories/Sale.java
@@ -39,7 +39,7 @@ public class Sale {
             stmt.setInt(QueryField.CUSTOMER_USER_ID, saleInfo.getCustomerUserId());            
             stmt.setInt(QueryField.STATUS_ID, saleInfo.getStatusId());            
             stmt.setInt(QueryField.SALE_DATE, saleInfo.getSaleDate());            
-            stmt.setLong(QueryField.DISCOUNT, saleInfo.getDiscount());
+            stmt.setDouble(QueryField.DISCOUNT, saleInfo.getDiscount());
             stmt.setString(QueryField.REMARKS, saleInfo.getRemarks());
             stmt.executeUpdate();
         }
@@ -57,8 +57,8 @@ public class Sale {
                 stmt.setInt(QueryField.PRODUCT_ID, productInfo.getId());
                 stmt.setString(QueryField.SALE_ORDER_NO, saleInfo.getOrderNo());
                 stmt.setString(QueryField.PURCHASE_ORDER_NO, productInfo.getPurchaseOrderNo());
-                stmt.setLong(QueryField.UNIT_PRICE, productInfo.getUnitPrice());
-                stmt.setLong(QueryField.DISCOUNT, productInfo.getDiscount());            
+                stmt.setDouble(QueryField.UNIT_PRICE, productInfo.getUnitPrice());
+                stmt.setDouble(QueryField.DISCOUNT, productInfo.getDiscount());            
                 stmt.executeUpdate();
             }
         }
@@ -74,8 +74,8 @@ public class Sale {
                 stmt.setString(QueryField.PURCHASE_ORDER_NO, productInfo.getPurchaseOrderNo());
                 stmt.setString(QueryField.SALE_ORDER_NO, saleInfo.getOrderNo());
                 stmt.setInt(QueryField.PRODUCT_ID, productInfo.getId());
-                stmt.setLong(QueryField.STOCK_OUT, productInfo.getQuantity());
-                stmt.setLong(QueryField.STOCK_IN, 0);
+                stmt.setDouble(QueryField.STOCK_OUT, productInfo.getQuantity());
+                stmt.setDouble(QueryField.STOCK_IN, 0);
                 //right now transaction category id constant. later update it from config file
                 stmt.setInt(QueryField.TRANSACTION_CATEGORY_ID, 5);           
                 stmt.executeUpdate();
